requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
category-encoders==2.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
deprecation==2.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastai==2.7.14
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.19.1
fastprogress==1.0.3
filelock==3.13.3
Flask==3.0.2
fonttools==4.51.0
fsspec==2024.3.1
idna==3.6
imbalanced-learn==0.12.2
importlib_metadata==7.1.0
ipython==8.23.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kaleido==0.2.1
kiwisolver==1.4.5
langcodes==3.3.0
lightgbm==4.3.0
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mpmath==1.3.0
murmurhash==1.0.10
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
orjson==3.10.0
packaging==24.0
pandas==2.1.4
parso==0.8.4
patsy==0.5.6
pillow==10.3.0
platformdirs==4.2.0
plotly==5.20.0
plotly-resampler==0.10.0
pmdarima==2.0.4
preshed==3.0.9
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycaret==3.3.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyod==1.1.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
retrying==1.3.4
rpds-py==0.18.0
schemdraw==0.15
scikit-base==0.7.5
scikit-learn==1.4.1.post1
scikit-plot==0.3.7
scipy==1.11.4
six==1.16.0
sktime==0.28.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.14.1
sympy==1.12
tbats==1.1.3
tenacity==8.2.3
thinc==8.2.3
threadpoolctl==3.4.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
traitlets==5.14.2
tsdownsample==0.1.3
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.2
widgetsnbextension==4.0.10
xxhash==3.4.1
yellowbrick==1.5
zipp==3.18.1
